+++ src/Models/EthereumWallet.php
@@ -23,6 +23,7 @@ class EthereumWallet extends Model
         'password',
         'mnemonic',
         'seed',
+        'derivation_path',
         'sync_at',
         'balance',
         'tokens',